+++ ui/mods/com.pa.quitch.ai-chat/live_game/messages.js
@@ -65,4 +65,5 @@ define({
   winning: ["!LOC:We're winning on"],
   losing: ["!LOC:We're losing on"],
   ok: ["!LOC:We're okay on"],
+  alone: ["!LOC:We own"],
 });